@@ -1,7 +1,8 @@
 package com.dji.sdk.sample.internal.view;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.StringRes;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.StringRes;
 
 /**
  * Basic required interface of a DemoView
