@@ -3,8 +3,6 @@ package com.dji.sdk.sample.demo.keymanager;
 import android.app.Service;
 import android.content.Context;
 import android.os.Handler;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.AdapterView;
@@ -15,8 +13,11 @@ import android.widget.Spinner;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.dji.sdk.sample.internal.view.PresentableView;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
 import com.dji.sdk.sample.R;
+import com.dji.sdk.sample.internal.view.PresentableView;
 
 import java.util.ArrayList;
 import java.util.HashMap;
